requirements.txt
bcrypt==3.1.6
cffi==1.12.3
Click==7.0
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.18.2
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug==0.15.4
Pandas
pyyaml

